@@ -21,7 +21,6 @@ classifiers =
     Programming Language :: Python
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -44,7 +43,7 @@ install_requires =
     virtualenv>=14.0.0
     importlib-metadata;python_version < '3.8'
     typing-extensions>=3.7.4;python_version < '3.8'
-python_requires = >=3.6
+python_requires = >=3.7
 include_package_data = True
 zip_safe = False
 
