@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Optional
+from __future__ import annotations
 
 from nox._options import noxfile_options as options
 from nox._parametrize import Param as param
@@ -20,6 +20,6 @@ from nox._parametrize import parametrize_decorator as parametrize
 from nox.registry import session_decorator as session
 from nox.sessions import Session
 
-needs_version: Optional[str] = None
+needs_version: str | None = None
 
 __all__ = ["needs_version", "parametrize", "param", "session", "options", "Session"]
